+++ RELEASE_NOTES.md
@@ -1,3 +1,12 @@
+### 0.12.2 March 31 2022 ####
+
+* Fix [deserialization type cache was shared between multiple serializer instances](https://github.com/akkadotnet/Hyperion/pull/306) 
+
+### 0.12.1 March 23 2022 ####
+
+* Fix disallow-unsafe-type Akka.NET settings and harden unsafe type detection [301](https://github.com/akkadotnet/Hyperion/pull/301)
+* Bump [Akka version from 1.4.34 to 1.4.35](https://github.com/akkadotnet/akka.net/releases/tag/1.4.35)
+
 ### 0.12.0 January 12 2022 ####
 
 * Allow explicit control over which types can be deserialized [#281](https://github.com/akkadotnet/Hyperion/pull/281)